ajustes e adição de funcionalidades

+++ Pedidex-ejb/src/main/java/com/jhonystein/pedidex/util/GenericDao.java
@@ -3,6 +3,7 @@ package com.jhonystein.pedidex.util;
 import com.jhonystein.pedidex.model.Entidade;
 import java.util.List;
 import javax.persistence.EntityManager;
+import javax.persistence.EntityNotFoundException;
 
 /**
  *
@@ -25,30 +26,44 @@ public class GenericDao<T extends Entidade> {
     }
 
     public T update(T bean) {
+        T oldBean = em.find(type, bean.getId());
+        if (oldBean == null) {
+            throw new EntityNotFoundException();
+        }
         return em.merge(bean);
     }
 
     public void delete(Long id) {
         T bean = em.getReference(type, id);
+        if (bean == null) {
+            throw new EntityNotFoundException();
+        }
         em.remove(bean);
     }
 
-    public List<T> findAll(Integer pageSize, Integer pageNumber) {
-        return this.findAll(pageSize, pageNumber, null, null, null);
+    private boolean nomEmpty(String... values) {
+        for (String val : values) {
+            if (val == null || val.trim().isEmpty()) {
+                return false;
+            }
+        }
+        return true;
     }
 
-    public List<T> findAll(Integer pageSize, Integer pageNumber, String filterField, String filterData) {
-        return this.findAll(pageSize, pageNumber, filterField, filterData, null);
+    public List<T> findAll(String filterField, String filterData, String order) {
+        return findAll(null, null, filterField, filterData, order);
     }
 
     public List<T> findAll(Integer pageSize, Integer pageNumber, String filterField, String filterData, String order) {
-        JpaCriteriaHelper helper = JpaCriteriaHelper.select(em, type)
-                .setPageSize(pageSize)
-                .page(pageNumber);
-        if (filterField != null && filterData != null) {
-            helper.where(filterField, JpaCriteriaHelper.ComparatorOperator.LIKE_IGNORE_CASE, filterData);
+        JpaCriteriaHelper helper = JpaCriteriaHelper.select(em, type);
+        if (pageSize != null && pageNumber != null) {
+            helper.setPageSize(pageSize)
+                    .page(pageNumber);
         }
-        if (order != null) {
+        if (nomEmpty(filterField, filterData)) {
+            helper.where(filterField, JpaCriteriaHelper.ComparatorOperator.LIKE_IGNORE_CASE, filterData.replaceAll("\\b", "%"));
+        }
+        if (nomEmpty(order)) {
             String[] parts = order.split("\\+");
             helper.orderBy(parts[0]);
             if (parts.length > 1 && parts[1].equalsIgnoreCase("desc")) {
@@ -59,7 +74,10 @@ public class GenericDao<T extends Entidade> {
     }
 
     public T find(Long id) {
-        return em.find(type, id);
+        T bean = em.find(type, id);
+        if (bean == null) {
+            throw new EntityNotFoundException();
+        }
+        return bean;
     }
-
 }
